Create owner policy authorization for adding and removing timeline

+++ src/Fletchling.Api/Controllers/TimelineController.cs
@@ -1,5 +1,6 @@
 ﻿using Fletchling.Api.Models;
 using Fletchling.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -7,20 +8,30 @@ using System.Threading.Tasks;
 
 namespace Fletchling.Api.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class TimelineController : ControllerBase
     {
+        private readonly IAuthorizationService _authService;
         private readonly ITimelineRepository _timelineRepo;
 
-        public TimelineController(ITimelineRepository timelineRepo)
+        public TimelineController(IAuthorizationService authService, ITimelineRepository timelineRepo)
         {
+            _authService = authService;
             _timelineRepo = timelineRepo;
         }
 
         [HttpPost]
         public async Task<IActionResult> AddTimelineToGroup([FromBody] TimelineRequest request)
         {
+            var authResult = await _authService.AuthorizeAsync(User, request, "OwnerPolicy");
+
+            if (!authResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             try
             {
                 await _timelineRepo.AddTimelineToGroup(request.UID, request.TwitterUsername, request.GroupName);
@@ -31,13 +42,20 @@ namespace Fletchling.Api.Controllers
                 Console.WriteLine(ex);
                 return StatusCode(StatusCodes.Status500InternalServerError);
             }
-            
+
             return Ok();
         }
 
         [HttpDelete]
         public async Task<IActionResult> RemoveTimelineFromGroup([FromBody] TimelineRequest request)
         {
+            var authResult = await _authService.AuthorizeAsync(User, request, "OwnerPolicy");
+
+            if (!authResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             try
             {
                 await _timelineRepo.RemoveTimelineFromGroup(request.UID, request.TwitterUsername, request.GroupName);
